+++ src/resources/privacy/privacy.api.ts
@@ -12,6 +12,7 @@ import {
   IgnorePrivacyPostParams,
   SetPrivacyPostParams,
   PrivacyAPI,
+  PiiTypeForColumnsResponse,
 } from './privacy.api.types';
 
 const privacyApiRequestArgsCreator = (configuration: Configuration): PrivacyAPIRequestArgsCreator => {
@@ -95,13 +96,13 @@ const privacyApi = (configuration: Configuration): PrivacyAPI => {
     async getPiiTypeForColumns(
       params: PiiTypeForColumnsParams,
       options?: AxiosRequestConfig
-    ): Promise<(axios: AxiosInstance, basePath: string) => AxiosPromise<string>> {
+    ): Promise<(axios: AxiosInstance, basePath: string) => AxiosPromise<PiiTypeForColumnsResponse[]>> {
       const requestAxiosArgs = await privacyApiRequestArgsCreator(configuration).getGetPiiTypeForColumnsRequestArgs(
         params,
         options
       );
 
-      return (axios: AxiosInstance = globalAxios, basePath: string): AxiosPromise<string> => {
+      return (axios: AxiosInstance = globalAxios, basePath: string): AxiosPromise<PiiTypeForColumnsResponse[]> => {
         const axiosRequestArgs = { ...requestAxiosArgs.options, url: basePath + requestAxiosArgs.url };
 
         return axios.request(axiosRequestArgs);