+++ src/resources/privacy/privacy.api.types.ts
@@ -11,6 +11,8 @@ export type IgnorePrivacyPostParams =
   paths['/api/Privacy/ignore']['post']['requestBody']['content']['application/json'];
 export type SetPrivacyPostParams = paths['/api/Privacy/set']['post']['requestBody']['content']['application/json'];
 
+export type PiiTypeForColumnsResponse = { table: string; column: string; piiType: string; }
+
 type PrivacyAPIRequestArgsCreator = {
   getGetPrivacyHistoryRequestArgs: (params: PrivacyHistoryParams, options?: AxiosRequestConfig) => Promise<RequestArgs>;
   getGetPrivacySuggestionsRequestArgs: (
@@ -45,7 +47,7 @@ type PrivacyAPI = {
   getPiiTypeForColumns: (
     params: PiiTypeForColumnsParams,
     options?: AxiosRequestConfig
-  ) => Promise<(axios: AxiosInstance, basePath: string) => AxiosPromise<string>>;
+  ) => Promise<(axios: AxiosInstance, basePath: string) => AxiosPromise<PiiTypeForColumnsResponse[]>>;
   ignorePrivacy: (
     params: IgnorePrivacyPostParams,
     options?: AxiosRequestConfig